+++ volmdlr/edges.py
@@ -20,7 +20,7 @@ import scipy.integrate as scipy_integrate
 from scipy.optimize import least_squares
 from geomdl import NURBS, BSpline
 
-from volmdlr.nurbs.operations import split_curve
+from volmdlr.nurbs.operations import split_curve, decompose_curve
 from volmdlr.nurbs.core import evaluate_curve, derivatives_curve
 from volmdlr.nurbs import fitting
 import volmdlr.nurbs.helpers as nurbs_helpers
@@ -1089,6 +1089,15 @@ class BSplineCurve(Edge):
         dict_['weights'] = self.weights
         return dict_
 
+    def decompose(self, return_params: bool = False):
+        """
+        Decomposes the curve into Bézier curve segments of the same degree.
+
+        :return: a list of Bezier segments
+        :rtype: list
+        """
+        return decompose_curve(self, return_params)
+
     def evaluate(self, **kwargs):
         """
         Evaluates the curve.
@@ -1192,7 +1201,7 @@ class BSplineCurve(Edge):
             return [None, self.copy()]
         if split_point.is_close(self.end, tol):
             return [self.copy(), None]
-        parameter = round(self.point_to_parameter(split_point), 12)
+        parameter = round(self.point_to_parameter(split_point), 19)
         return split_curve(self, parameter)
 
     def get_reverse(self):
@@ -1332,6 +1341,8 @@ class BSplineCurve(Edge):
 
         :return: the given point when the BSplineCurve3D is evaluated at the u value.
         """
+        if self.length() == 0.0:
+            print(True)
         u = max(min(abscissa / self.length(), 1.), 0.0)
         u_min, u_max = self.domain
         if u_min != 0 or u_max != 1.0:
@@ -5047,11 +5058,8 @@ class BSplineCurve3D(BSplineCurve):
 
         knot_multiplicities = [int(i) for i in arguments[6][1:-1].split(",")]
         knots = [float(i) for i in arguments[7][1:-1].split(",")]
-        knot_vector = []
-        for i, knot in enumerate(knots):
-            knot_vector.extend([knot] * knot_multiplicities[i])
 
-        if 9 in range(len(arguments)):
+        if len(arguments) >= 10:
             weight_data = [float(i) for i in arguments[9][1:-1].split(",")]
         else:
             weight_data = None
@@ -5239,7 +5247,7 @@ class BSplineCurve3D(BSplineCurve):
         if self.end.is_close(point3d):
             return self.reverse()
 
-        curves = volmdlr.nurbs.operations.split_curve(self, round(parameter, 7))
+        curves = volmdlr.nurbs.operations.split_curve(self, round(parameter, 19))
         return curves[1]
 
     def cut_after(self, parameter: float):
@@ -5255,7 +5263,7 @@ class BSplineCurve3D(BSplineCurve):
             return self.reverse()
         if self.end.is_close(point3d):
             return self.copy()
-        curves = volmdlr.nurbs.operations.split_curve(self, round(parameter, 7))
+        curves = volmdlr.nurbs.operations.split_curve(self, round(parameter, 19))
         return curves[0]
 
     def insert_knot(self, knot: float, num: int = 1):